feat: Softmax, mat concat column, and more impl and tested

+++ src/constants.h
@@ -10,6 +10,6 @@
 #define MAX_SEQ_LEN 64               // words
 #define D_MODEL 64                   // dims
 #define NUM_HEADS 8                  // words
-#define D_HEAD (D_MODEL / NUM_HEADS) // head dim
+#define D_HEAD ((double)D_MODEL / NUM_HEADS) // head dim
 
 #endif
